renamed folder to GameComponents as more components will be made
